refactor: move brc home sectionhero to sites and reuse shared hero (#1598)
