--- app/components/Home/components/Section/components/SectionHero/components/Hero/hero.styles.ts
@@ -1,6 +1,0 @@
-import styled from "@emotion/styled";
-
-export const SVG = styled.svg`
-  position: absolute;
-  z-index: -1;
-`;
