--- app/components/Home/components/Section/components/SectionHero/constants.ts
@@ -1,7 +1,0 @@
-import { type ButtonProps } from "@mui/material";
-
-export const BUTTON_PROPS: Partial<ButtonProps> = {
-  color: "primary",
-  size: "large",
-  variant: "contained",
-};
