+++ app/views/HomeView/homeView.tsx
@@ -1,8 +1,8 @@
 import { SectionAnalytics } from "@/components/Home/components/Section/components/SectionAnalytics/sectionAnalytics";
 import { SectionAssemblies } from "@/components/Home/components/Section/components/SectionAssemblies/sectionAssemblies";
-import { SectionHero } from "@/components/Home/components/Section/components/SectionHero/sectionHero";
 import { SectionSubHero } from "@/components/Home/components/Section/components/SectionSubHero/sectionSubHero";
 import { SectionWhitePapers } from "@/components/Home/components/Section/components/SectionWhitePapers/sectionWhitePapers";
+import { SectionHero } from "@brc/views/HomeView/components/SectionHero/sectionHero";
 import { Fragment, type JSX } from "react";
 
 export const HomeView = (): JSX.Element => {
