+++ sites/brc-analytics/views/HomeView/components/SectionHero/components/Carousel/cards/constants.ts
@@ -1,4 +1,4 @@
-import * as MDX from "@/components/Home/components/Section/components/SectionHero/components/Carousel/content";
+import * as MDX from "@brc/views/HomeView/components/SectionHero/components/Carousel/content";
 import { type CardProps } from "@databiosphere/findable-ui/lib/components/common/Card/card";
 
 export const CAROUSEL_CARDS: Pick<CardProps, "text">[] = [
