+++ sites/brc-analytics/views/HomeView/components/SectionHero/components/Carousel/carousel.styles.ts
