+++ sites/brc-analytics/views/HomeView/components/SectionHero/components/Carousel/carousel.tsx
