+++ sites/brc-analytics/views/HomeView/components/SectionHero/components/Carousel/common/constants.ts
