+++ sites/brc-analytics/views/HomeView/components/SectionHero/components/Carousel/common/utils.ts
