+++ sites/brc-analytics/views/HomeView/components/SectionHero/components/Carousel/components/Arrow/arrow.styles.ts
@@ -1,5 +1,5 @@
-import { MAX_DECK_SIZE } from "@/components/Home/components/Section/components/SectionHero/components/Carousel/common/constants";
-import { getArrowTransform } from "@/components/Home/components/Section/components/SectionHero/components/Carousel/common/utils";
+import { MAX_DECK_SIZE } from "@brc/views/HomeView/components/SectionHero/components/Carousel/common/constants";
+import { getArrowTransform } from "@brc/views/HomeView/components/SectionHero/components/Carousel/common/utils";
 import { COLOR_MIXES } from "@databiosphere/findable-ui/lib/styles/common/constants/colorMixes";
 import { PALETTE } from "@databiosphere/findable-ui/lib/styles/common/constants/palette";
 import { bpDownSm } from "@databiosphere/findable-ui/lib/styles/common/mixins/breakpoints";
