+++ sites/brc-analytics/views/HomeView/components/SectionHero/components/Carousel/components/Arrow/arrow.tsx
