+++ sites/brc-analytics/views/HomeView/components/SectionHero/components/Carousel/components/Cards/cards.styles.ts
@@ -2,12 +2,12 @@ import {
   MAX_CARD_HEIGHT,
   MAX_CARD_HEIGHT_SM,
   MAX_CARD_WIDTH,
-} from "@/components/Home/components/Section/components/SectionHero/components/Carousel/common/constants";
+} from "@brc/views/HomeView/components/SectionHero/components/Carousel/common/constants";
 import {
   getCardTransform,
   getCardTransition,
   getCardZIndex,
-} from "@/components/Home/components/Section/components/SectionHero/components/Carousel/common/utils";
+} from "@brc/views/HomeView/components/SectionHero/components/Carousel/common/utils";
 import { FONT } from "@databiosphere/findable-ui/lib/styles/common/constants/font";
 import { PALETTE } from "@databiosphere/findable-ui/lib/styles/common/constants/palette";
 import { SHADOWS } from "@databiosphere/findable-ui/lib/styles/common/constants/shadows";
