+++ sites/brc-analytics/views/HomeView/components/SectionHero/components/Carousel/content/aiCodeathon2.mdx
