+++ sites/brc-analytics/views/HomeView/components/SectionHero/components/Carousel/content/genomeArk2.mdx
