+++ sites/brc-analytics/views/HomeView/components/SectionHero/components/Carousel/content/userSurvey.mdx
