+++ sites/brc-analytics/views/HomeView/components/SectionHero/components/Carousel/hooks/types.ts
@@ -1,0 +1,11 @@
+import { type CardProps } from "@databiosphere/findable-ui/lib/components/common/Card/card";
+import { type UseSwipeInteraction } from "@repo/shared/views/HomeView/hooks/UseSwipeInteraction/types";
+
+export interface UseInteractiveCarousel {
+  activeIndex: UseSwipeInteraction["activeIndex"];
+  interactiveAction?: UseSwipeInteraction["interactiveAction"];
+  interactiveCards: CardProps[];
+  interactiveIndexes: number[];
+  onSetActiveIndex: UseSwipeInteraction["onSetActiveIndex"];
+  onSetSwipeAction: UseSwipeInteraction["onSetSwipeAction"];
+}
