+++ sites/brc-analytics/views/HomeView/components/SectionHero/sectionHero.styles.ts
